backend/requirements.txt
asgiref
certifi
cffi
charset-normalizer
coreapi
coreschema
cryptography
defusedxml
Django
django-cors-headers
django-templated-mail
djangorestframework
djangorestframework-simplejwt
djoser
idna
itypes
Jinja2
MarkupSafe
oauthlib
Pillow
psycopg2
psycopg2-binary
psycopg2-pool
pycparser
PyJWT
python-decouple
python-dotenv
python3-openid
pytz
requests
requests-oauthlib
six
social-auth-app-django
social-auth-core
sqlparse
tzdata
uritemplate
urllib3
